Added authentication, Redis integration, and error handling.

+++ backend/BankCardProject/Controllers/BankCardsController.cs
@@ -1,8 +1,11 @@
 ﻿using AutoMapper;
 using BankCardProject.DTOs;
+using BankCardProject.Exceptions;
 using BankCardProject.Models;
+using BankCardProject.Properties;
 using BankCardProject.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
+using System.Runtime.Versioning;
 
 namespace BankCardProject.Controllers
 {
@@ -11,12 +14,10 @@ namespace BankCardProject.Controllers
     public class BankCardsController : ControllerBase
     {
         private readonly IBankCardService _service;
-        private readonly IMapper _mapper;
 
-        public BankCardsController(IBankCardService service, IMapper mapper)
+        public BankCardsController(IBankCardService service)
         {
             _service = service;
-            _mapper = mapper;
         }
 
         /// <summary>
@@ -26,8 +27,6 @@ namespace BankCardProject.Controllers
         public async Task<ActionResult<List<BankCardDto>>> GetAll()
         {
             var dto = await _service.GetAllCardsAsync();
-            if (dto == null)
-                return NotFound("Banka kartı listesi boş.");
             return Ok(dto);
         }
 
@@ -37,10 +36,14 @@ namespace BankCardProject.Controllers
         [HttpGet("byId/{id}")]
         public async Task<ActionResult<BankCardDto>> GetById(int id)
         {
+            if (id <= 0)
+            {
+                throw new InvalidParameterException(Resources.ERR1015);
+            }
             var dto = await _service.GetCardByIdAsync(id);
 
             if (dto == null)
-                return NotFound("Id ile eşleşen kart bilgisi bulunamadı.");
+                throw new NotFoundException(Resources.CRUD2001);
 
             return Ok(dto);
         }
@@ -49,17 +52,14 @@ namespace BankCardProject.Controllers
         /// Yeni bir BankCard oluşturur.
         /// </summary>
         [HttpPost("create")]
-        public async Task<ActionResult<BankCardDto>> Create(BankCardDto dto)
+        public async Task<IActionResult> Create(BankCardDto dto)
         {
             if (dto == null)
-                return BadRequest("BankCard bilgileri boş olamaz.");
+                throw new BadRequestException(Resources.CRUD1002);
 
-            bool isCreated = await _service.CreateCardAsync(dto);
+            await _service.CreateCardAsync(dto);
 
-            if (!isCreated)
-                return BadRequest("BankCard oluşturulamadı.");
-
-            return Ok(dto);
+            return Ok(Resources.CRUD1000);
         }
 
         /// <summary>
@@ -68,16 +68,17 @@ namespace BankCardProject.Controllers
         [HttpPut("updateById/{id}")]
         public async Task<IActionResult> Update(int id, BankCardDto cardDto)
         {
+            if (id <= 0)
+            {
+                throw new InvalidParameterException(Resources.ERR1015);
+            }
             if (cardDto == null)
             {
-                return BadRequest(new { Message = "Güncelleme verisi sağlanmalıdır." });
+                throw new BadRequestException(Resources.CRUD3002);
             }
-            bool isUpdated = await _service.UpdateCardAsync(id, cardDto);
+            await _service.UpdateCardAsync(id, cardDto);
 
-            if (!isUpdated)
-                return BadRequest("BankCard güncellenemedi.");
-
-            return Ok("BankCard başarıyla güncellendi.");
+            return Ok(Resources.CRUD3000);
         }
 
         /// <summary>
@@ -86,11 +87,13 @@ namespace BankCardProject.Controllers
         [HttpDelete("deleteById/{id}")]
         public async Task<IActionResult> Delete(int id)
         {
-            bool isDeleted = await _service.DeleteCardAsync(id);
-            if (!isDeleted)
-                return BadRequest("BankCard silinemedi.");
-
-            return Ok(isDeleted);
+            if (id <= 0)
+            {
+                throw new InvalidParameterException(Resources.ERR1015);
+            }
+            await _service.DeleteCardAsync(id);
+     
+            return Ok(Resources.CRUD4000);
         }
     }
 }
